Fix functionality of LanguageSwitcher component
Tricking typescript into being normal. Honestly I get cpp being this picky, but this is unreasonable. Added detailed type definition and the function getAutonym to handle missing local language names in the LanguageSwitcher component. This enables the component to display the available language options in a more precise way and handle a wider range of languages, enhancing the UX on multilingual sites.