Add JWT callback to auth options for setting accessToken
The callback 'jwt' was added to the authentication options in the '[...nextauth].ts' file. This callback sets the 'accessToken' on the token object, which gets created during the authentication process. It allows the access token to be stored and used in frontend for subsequent API calls. Multiple access token sources are supported and it accounts for both 'accessToken' and 'access_token' formats.