Refactor chat handleSend code to improve readability
Code was really messy and difficult to parse, though not necessarily poorly written. This was a prerequisite to enabling apim as a backend. Split the handleSend function from the chat component into several smaller helper functions. This makes the code easier to read and maintain. These new functions include updateConversationFromUserInput, makeRequest, setConversationTitle, handleGoogleResponse, handleNormalChatBackendStreaming which abstract the steps in sending the chat data to the server and updating the UI accordingly.