Add APIM endpoints to constants with defaults
The constants APIM_CHAT_ENDPONT and APIM_MANAGEMENT_ENDPONT have been added to const.ts. These changes allow default fallback values to be specified for the application program interface management (APIM) chat and management endpoints when they are not set in the environment variables. It also supports changing these through env variables if we so decide.