Handle 400 error in APIM server
Added an error handler for when the APIM server returns a 400 status, specifically when the 'content_length_exceeded' code is given. It will now return a response with the same status, headers, and body as the original response from the server. An instance of APIMError will also be thrown, containing details from the JSON object returned by the server. Additionally, the error message for non-200 statuses has been updated to include the text and status of the response.