Refactor how pluginKeys are parsed from HomeContext
The `pluginKeys` from `Chat.tsx`, `PluginKeys.tsx` and `Chatbar.tsx` are now checked if they're a string and, if true, parsed into a JSON object. The implementation changed because previously `pluginKeys` were assumed to be a JSON object which caused errors if not valid JSON.

Additionally, a bug was fixed in the logic for conversations in `Chat.tsx` to improve the conversation handling process. Now, `setConversationTitle` function is only called when `updatedConversation` contains only one message, and also current conversation is now properly updated with `updatedConversation` variable. Previously it was using `updateConversation`, which was a function and preventing the conversation from actually updating

Lastly, a null-check was added in `Chat.tsx` to handle cases when model name does not exist for selected conversation.