Reimplement default value for plugin in handleSend method
In the handleSend method within Chat.tsx, the parameter 'plugin' now has a default value of null when not supplied. This change provides further defense against a potential undefined value for plugin, simplifying code and reducing potential for errors. It used to work this way, but was changed when messing with types. So reimplementing.