Add forced logout on API call failure, optional to disable
Code was added in errorService for getting a logout when API calls fail (implicitly model api call, which is mostly a backend auth failure). The access to a new variable, FORCE_LOGOUT_ON_REFRESH_FAILURE, was added in the constants file. This helps the issue of inactive sessions becoming 500 errors that are not user friendly.