Add previously unsupported OpenAI models and add default values
This update expands the list of OpenAI models enumerated in the code, while also extending the default parameters for several of these models. Particularly, the additions include details for models like GPT-3.5-16k, GPT-4-Turbo and GPT-4-Vision. GPT-4-Vision has been commented out even though it can work as a simple chat model. This is because its inclusion would probably confuse given that the platform doesn't currently support image interactions in any way, shape, or form. The default values are guesses and for GPT-4-Turbo we don't actually have enough tokens overally to use the model as is. But that's the maximum, so I'm keeping it like that for now.