Add in a new model, better distinguishing between versions of gpt-4
The code now includes a new addition - the 'GPT_4_0613' model. This was done in order to make a clear distinction between 'GPT_4' and 'GPT_4_0613' versions. Understandably, the differentiation between models will make it easier to identify and utilize each version based on their unique properties. Also, each version maintains a maxLength of 24000 and a tokenLimit of 8000, as previously defined.