Extend Message interface with enum to support new functionality types
A MessageType enum has been created to differentiate message types (text, image, audio, video). This enum has been added as a new property to the Message interface to add flexibility and specificity to chat message content management. Undefined is left as an option for messageType, which should default to TEXT and minimize disruption to existing features.