Fix handling of blob factory, fix missed type bug on getting blobs, and introduce enum for factory
The code has been revised to throw appropriate errors when a blob download fails or when a blob isn't found. BlobStorageType, a new enum, has been introduced to handle the creation of different types of blob storage. BlobStorageFactory has been updated to use this new enum, providing more flexibility in the creation of blob storage and enhancing the robustness of the code.