Fix type issues after introduction of messageType, minor imports formatting change
In this commit, messageType is added to the Message structure, ensuring all Message instances now include a messageType field. This is done to fix type mismatches that arose after messageType was introduced in the Message type. While doing that, some adjustments have been made to the formatting of imports to make the code cleaner and more readable.