Add getEnvVariable utility function
This new function is designed to fetch environment variables. It takes the name of the environment variable, a boolean indicating whether to throw an error if the variable doesn't exist, and a default value to return if no error is thrown and the variable doesn't exist. This simplifies the logic for handling env variables in other areas of the application while also keeping stricter typing requirements.