Add createContainer method to blob service
A new method, createContainer, has been added to the blob service to allow the creation of a new container. This change provides more flexibility enabling us to create a container directly through blob service client. Not currently implemented anywhere in the application, as the use case is better handled in IaC, but keeping for future potential uses.