Implement image upload function in API
The implementation of the image upload function has been accomplished in the page's api/image.ts. With the help of AzureBlobStorage client, images get uploaded to the desired container. Hashing has also been introduced to create a unique name for each file. Upon successful upload, the function sets HTTP status to 200 and responds with the URI of the uploaded image.