Add new helper function to manage chat messages
A new helper function named "getMessagesToSend" was added to manage chat messages. This function calculates the number of tokens in a message and checks if it exceeds the token limit before sending. Message management code was refactored in the api chat.ts file to use this new function. This is a prerequisite to handling different types of chat messages as well.