Add image fetching in ChatMessageImage component
Jesus Christ this would've taken me 5 minutes in python. Added useState and useEffect hooks to fetch an image from a given URL and convert it to a base64 string in ChatMessageImage component. This change helps to replace the direct image URL source in the img tag with the base64 format of the image. For some reason this was the only way to make this work. Javascript is actually torture to use.