Refactored `getMessagesToSend` to handle async operations
The 'getMessagesToSend' function in chat utility has been updated to handle asynchronous operations. A Promise has been introduced along with adjustments to handle message contents that are arrays. Also added was support for extracting a Base64 encoded image from a URL if a contentSection type is identified as an image. This was all done to better match the structure and style of requests to the vision endpoint.