Add file preview components
New components have been added to preview file uploads to the chat. The user now has the ability to see file uploads and remove them before submitting. A new "XIcon" component was also created for the remove button in the interface. It is not fully hooked up, even internally, but the basic infrastructure is here.