Update blob upload method to handle array content
The blob uploading function in the server utility has been updated to handle content that comes in an array form. It now checks if the provided content is an array and, if so, safely assigns the first element of the array as the content to be uploaded. The change attempts to fix a recurrent CodeQL error in github that seems to be a false positive but actually provided a helpful message this last time.