Add utility function for getting full message from prompt and content
This commit adds a new function, getFullMessageFromPromptAndContent, in the 'prompts.ts' utility file. This function returns a full message composed of prompt content and user-defined content which can be either a string or an array of TextMessageContent or ImageMessageContent. This function will be useful for generating complete messages with dynamic content. It has not been implemented yet because it is beyond the current scope of needs, so it is untested.