Add submit type state to ChatFileUploadPreviews
The ChatFileUploadPreviews component and related properties have been updated to include a 'setSubmitType' property, that's dispatched when a file upload preview is removed. When there are no more file previews, the submit type is changed to 'text'. Also, minor visual adjustments have been made to the file preview removal button for improved UI.