Add function to check if conversation is image-based
A new utility function has been added to the application chat utils to check whether a conversation is image-based or not. It does so by checking the type of the last message in a pool of messages. If the last message content includes an 'image_url', it is considered an image conversation.