Add getMessagesToSendV2 function in chat utils
A new function, getMessagesToSendV2, has been added in the chat utilities. This function improves the previous getMessagesToSend function by handling images in the conversation, translating them to appropriate string format. It meticulously processes both textual and image-based content, thereby enhancing the functionality of the chat application. The old method was preserved because this was a significant refactor, but will be removed eventually.