Move 'checkIsModelValid' function to 'utils/app/chat.ts'
The 'checkIsModelValid' function has been moved from 'pages/api/chat.ts' to 'utils/app/chat.ts' to improve maintainability. This function checks the validity of a given model ID based on a set of valid model IDs. This separates concerns better and allows reuse in other areas of the application.