Refactor image capture function and enhance UI interactions
This update includes a refactoring of the image capture function to return a Promise. Additional state actions were passed to onTakePhotoButtonClick for enhanced usability. Minor changes were also made to UI elements including adding class names to button elements. Currently the implementation is working, but the display for taking the photo is bad looking.