Upload files in chunks to support larger files
Previously, files were uploaded in a single request. Now, files are uploaded in 5MB chunks. We have not changed the 10mb limit yet. This allows supporting larger files while avoiding timeouts or memory issues.

The file upload endpoint remains the same, but now expects a "chunk" of the file and the file name in the FormData. The frontend repeatedly sends chunks until the entire file is uploaded. Toasts provide user-facing success/error messages, though some of these will be removed as functionality is fleshed out.