Remove code unnecessary to fetching file from Azure Blob Storage
Remove unused SAS token generation code. Directly fetch the file from Azure Blob Storage using the blob storage client and return it in the API response.

The SAS token generation code is removed as it is no longer needed. The file is now fetched directly using the `blobStorageClient.get()` method, passing the blob location and `BlobProperty.BLOB` to retrieve the file content as a
Buffer. The fetched file is then returned in the API response. The value is already base64, so no need to have a `toString('base64')`.