Fix handling of submit type with file uploads
This update revises the file upload mechanisms within the chat module. Rather than setting the submit type within the file upload event dispatcher on the frontend, the submit type setting functionality has been moved to the backend within the onFileUpload function. This ensures a more consistent and accurate update of the submit type, contingent on the success or failure of the file upload.

Hopefully this won't be necessary in production, as file uploads shouldn't fail if they get to that stage. But it is useful for development and won't hurt as a fallback for production.