Add support for file message type in chat
The ChatInput component in the Chat module has been updated to handle 'file' as a type of message content. This allows users to input a file URL when sending messages. Now, Message objects created on form submission can have 'file' or 'image' as messageType instead of just 'image' or 'text'.