Add blobType parameter to getBlobBase64String function
The function getBlobBase64String in server/blob.ts has been updated to include a new parameter 'blobType'. This parameter, with a default value of 'images', allows the function to handle different types of blobs, such as files, images, audio, and video. The blob location string has been updated to reflect this change.