Add messageType property to handleSend function
The changes in this commit involve adding a messageType property to the 'handleSend' function in the 'Chat.tsx' component. This property is necessary to specify the type of message when sending chat messages, in this case, it's a "text" type. Several unnecessary imports were also removed in the process. This is cleanup from after fixing the explicit merge conflicts.